agent1/pyproject.toml
[project]
name = "365"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-dotenv",
    "openai",
    "requests",
    "serpapi",
    "cachetools",
    "streamlit",
    "serpapi",
    "python-telegram-bot",
]
